Move the help topics around.
